% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generations.R
\name{apc_ef_vcov}
\alias{apc_ef_vcov}
\title{Covariance of an estimable function}
\usage{
apc_ef_vcov(M, comp = c("lac", "cac", "ftt", "fcp", "ld"), ref = NULL)
}
\arguments{
\item{M}{fitted model (\code{\link[=apc_fit]{apc_fit()}}).}

\item{comp}{estimable function.}

\item{ref}{reference values (see \code{\link[=apc_ref]{apc_ref()}}).}
}
\value{
list with \code{x}, \code{est} (log scale) and covariance \code{V}.
}
\description{
Covariance of an estimable function
}
