% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apc.R
\name{apc_scale}
\alias{apc_scale}
\title{Scale of an estimable function: log scale to rate per `per` (basic EFs) or annual \% change (ld)}
\usage{
apc_scale(comp, per = 1e+05)
}
\arguments{
\item{comp}{estimable function.}

\item{per}{rate multiplier.}
}
\value{
a function.
}
\description{
Scale of an estimable function: log scale to rate per `per` (basic EFs) or annual \% change (ld)
}
