% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{pchip}
\alias{pchip}
\title{MATLAB pchip interpolation}
\usage{
pchip(x, y, xi)
}
\arguments{
\item{x, y}{knots (x increasing).}

\item{xi}{points to interpolate.}
}
\value{
interpolated values.
}
\description{
Shape-preserving piecewise cubic Hermite interpolation with MATLAB's slopes (pchipslopes).
}
