% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{rates_uni}
\alias{rates_uni}
\title{Summary statistics of a rates object (cf. MATISSE uni)}
\usage{
rates_uni(R)
}
\arguments{
\item{R}{rates object.}
}
\value{
one-row data.table.
}
\description{
Summary statistics of a rates object (cf. MATISSE uni)
}
