% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generations.R
\name{social_generations}
\alias{social_generations}
\title{Social generations and proxy parents (JAMA Netw Open 2024)}
\usage{
social_generations()
}
\value{
data.table: \code{generation}, first and last birth year \code{y0}, \code{y1}.
}
\description{
Social generations and proxy parents (JAMA Netw Open 2024)
}
