% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apc.R
\name{apc_efs}
\alias{apc_efs}
\title{Several estimable functions, long format}
\usage{
apc_efs(M, comp = c("lac", "cac", "ftt", "fcp", "ld"), ref = NULL, ...)
}
\arguments{
\item{M}{fitted model ([apc_fit()]).}

\item{comp}{estimable function.}

\item{ref}{reference values (see [apc_ref()]).}

\item{...}{passed to [apc_ef()].}
}
\description{
Several estimable functions, long format
}
