% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capricorn.R
\name{cap_blocks}
\alias{cap_blocks}
\title{Coefficient blocks of the APC design (GEODE pointers)}
\usage{
cap_blocks(G)
}
\arguments{
\item{G}{design (\code{\link[=apc_design]{apc_design()}}).}
}
\description{
1 intercept, 2 LAT, 3 net drift, 4-6 quadratic age/period/cohort, 7-9 higher-order age/period/cohort deviations.
}
