% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matisseR-package.R
\docType{package}
\name{matisseR-package}
\alias{matisseR}
\alias{matisseR-package}
\title{matisseR: age-period-cohort analysis of vital rates, ported from MATISSE}
\description{
An R port of methods in the MATISSE MATLAB toolbox (Philip S. Rosenberg, US National Cancer
Institute): Lexis diagrams of rates (`rates_*`, [csv2rates()], [rates_i5()]), the New
Age-Period-Cohort Model and its estimable functions ([apc_fit()], [apc_ef()]) and hypothesis-based
comparative APC analysis ([capricorn()], [cap_ef()]). The MATISSE example datasets are included
([matisse_example()]).
}
\references{
Miranda Filho A, Rosenberg PS. Advances in statistical methods for cancer surveillance
  research: an age-period-cohort perspective. Front Oncol 2024;13:1332429. doi:10.3389/fonc.2023.1332429
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/filhoalm/matisseR}
  \item Report bugs at \url{https://github.com/filhoalm/matisseR/issues}
}

}
\author{
\strong{Maintainer}: Adalberto Miranda Filho \email{mirandaa@iarc.who.int}

Other contributors:
\itemize{
  \item Philip S. Rosenberg (author of the MATISSE MATLAB toolbox and its methods) [contributor]
}

}
\keyword{internal}
